Iteration
Iteration is a process of using a statement to loop the function until the statement becomes false. It is a really nice technique which can make various mathematical functions and  can create a really nice setting for higher calculations.
Notes : 1. Very often the final value of the entered variable goes to zero as here is a case when we use print(n) it gives 0 as result but we wanted to print n! = fact(n) which can be done just by using another variable with value (say k) = 0 and use the operation k=k+1 so that it finally provides the value of n to us which is now zero but k is n.
2. In iteration we use the operations in order always keep the answer operation at the top to be honest.