cube root for an integer problem by looping
The very obvious algorithm for this program can be understood through analysing what we need to do for finding the cube root of any integer.  We need to find a way so that the program gives the positive value for cube roots for positive integer and negative value for negative integer and also tells us if the integer that we provided is not a perfect cube if it can't find it.
To pull this of we approach the value of the given integer x step by step by adding 1 to our answer and finally reach to our cube root. If on approaching the integer value with  the cube root we find that it's cube is not equal to abs(x) we want it to tell us that it's not perfect and for negative value of x we must change the sign of the root as we were approaching positive abs(x).
Thus the algorithm is:
1.  Ask for the integer from the user and declare a variable which will be our answer to be zero.
2. Delay the the process of adding 1 to our answer variable by looping the truth value till it become false ( iteration ) and by considering the statement that our answer cube must be less than the abs(x).
3. Use if statement to print if answer cube is not equal to abs(x).
4. Use else statement to print answer if x is a perfect cube also in else we nest an if statement to change the sign of our answer if x is less than zero.