Algorithm for the problem
The program initializes the variable found_pair to False, root and power to 1. The first while loop checks if root raised to the power is less than or equal to the integer entered by the user and power is less than 6. If it is, it enters the inner while loop, which checks if root raised to the power is equal to the integer. If it is, it sets found_pair to True and breaks out of the inner while loop. If it does not find such a pair, it increments power by 1 and continues the inner while loop until either it finds a pair or power reaches 6. If it finds a pair, it breaks out of the outer while loop as well. If it does not find a pair, it increments root by 1 and resets power to 1 and continues the outer while loop until either it finds a pair or root ** power exceeds the integer. Once it has found a pair or completed both the loops, it checks the value of found_pair. If it is True, it prints the root and power that satisfy the equation. Otherwise, it prints a message saying that no such pair exists.