Doing the power root problem with for loops only
The program initializes the variable found_pair to False. It uses two for loops to check for the root and power values that satisfy the equation. The outer for loop iterates through all values of root from 1 to integer - 1. The inner for loop iterates through all values of power from 1 to 5. If root raised to the power is equal to the integer entered by the user, the program sets found_pair to True and breaks out of both loops. If it does not find such a pair, it continues the loops until either it finds a pair or it has checked all possible combinations of root and power. Once it has found a pair or completed both loops, it checks the value of found_pair. If it is True, it prints the root and power that satisfy the equation. Otherwise, it prints a message saying that no such pair exists.